Refactor SetupWizardRoute
