+++ client/components/setupWizard/SetupWizardPage.css
@@ -1,0 +1,42 @@
+.SetupWizard {
+	display: flex;
+
+	width: 100%;
+	height: 100vh;
+
+	background-color: var(--color-dark-05);
+	align-items: stretch;
+
+	justify-content: center;
+
+	&__wrapper {
+		flex: 1 1 auto;
+
+		height: 100vh;
+		padding: 1rem;
+	}
+
+	&__steps {
+		overflow: auto;
+
+		width: 100%;
+		height: 100%;
+		padding: 1rem 3rem;
+
+		border-radius: 2px;
+		background-color: var(--color-white);
+		box-shadow: 0 2px 4px 0 rgba(0, 0, 0, 0.08);
+		justify-content: center;
+	}
+}
+
+@media (width <= 760px) {
+	.SetupWizard {
+		flex-direction: column;
+		justify-content: initial;
+
+		&__wrapper {
+			padding: 0;
+		}
+	}
+}
