+++ client/components/setupWizard/SetupWizardPage.js
@@ -1,5 +1,4 @@
 import React from 'react';
-import styled from 'styled-components';
 
 import { useTranslation } from '../../contexts/TranslationContext';
 import { useWipeInitialPageLoading } from '../../hooks/useWipeInitialPageLoading';
@@ -11,49 +10,16 @@ import { AdminUserInformationStep } from './steps/AdminUserInformationStep';
 import { SettingsBasedStep } from './steps/SettingsBasedStep';
 import { RegisterServerStep } from './steps/RegisterServerStep';
 
-const Container = styled.div`
-	display: flex;
-	width: 100%;
-	height: 100vh;
-	background-color: var(--color-dark-05, #f1f2f4);
-	align-items: stretch;
-	justify-content: stretch;
-
-	@media (max-width: 760px) {
-		flex-direction: column;
-		justify-content: initial;
-	}
-`;
-
-const Wrapper = styled.section`
-	flex: 1 1 auto;
-	height: 100vh;
-	padding: 1rem;
-
-	@media (max-width: 760px) {
-		padding: 0;
-	}
-`;
-
-const Steps = styled.div`
-	overflow: auto;
-	width: 100%;
-	height: 100%;
-	padding: 1rem 3rem;
-	border-radius: 2px;
-	background-color: var(--color-white, white);
-	box-shadow: 0 2px 4px 0 rgba(0, 0, 0, 0.08);
-	justify-content: center;
-`;
+import './SetupWizardPage.css';
 
 export function SetupWizardPage() {
 	useWipeInitialPageLoading();
-	const { currentStep } = useSetupWizardStepsState();
+	const { currentStep = 1 } = useSetupWizardStepsState();
 	const t = useTranslation();
 
 	return <>
 		<ConnectionStatusAlert />
-		<Container>
+		<div className='SetupWizard'>
 			{(currentStep === finalStep && <Epilogue />)
 			|| <>
 				<SideBar
@@ -77,15 +43,15 @@ export function SetupWizardPage() {
 					]}
 					currentStep={currentStep}
 				/>
-				<Wrapper>
-					<Steps>
+				<div className='SetupWizard__wrapper'>
+					<section className='SetupWizard__steps'>
 						<AdminUserInformationStep step={1} title={t('Admin_Info')} active={currentStep === 1} />
 						<SettingsBasedStep step={2} title={t('Organization_Info')} active={currentStep === 2} />
 						<SettingsBasedStep step={3} title={t('Server_Info')} active={currentStep === 3} />
 						<RegisterServerStep step={4} title={t('Register_Server')} active={currentStep === 4} />
-					</Steps>
-				</Wrapper>
+					</section>
+				</div>
 			</>}
-		</Container>
+		</div>
 	</>;
 }
